Feat : 디테일 프로필 구현
