+++ AltoYou/FirstTabView/MainViewController.swift
@@ -1,0 +1,26 @@
+//
+//  MainViewController.swift
+//  AltoYou
+//
+//  Created by 정의찬 on 10/3/23.
+//
+
+import UIKit
+
+class MainViewController: UIViewController {
+    
+    // MARK: - init
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setSelf()
+    }
+    
+    // MARK: - Function
+    private func setSelf(){
+        self.view.backgroundColor = UIColor(red: 0.49, green: 0.64, blue: 1.00, alpha: 1.00)
+    }
+    
+    
+    // MARK: - Constraints
+
+}
