+++ AltoYou/SecondTabView/SecondMainViewController.swift
@@ -1,0 +1,29 @@
+//
+//  SecondMainViewController.swift
+//  AltoYou
+//
+//  Created by 정의찬 on 10/4/23.
+//
+
+import UIKit
+
+class SecondMainViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        self.view.backgroundColor = .white
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
